cordova: copyLibs adds all dependencies from projects
To update the dependencies in the cordova plugin, run :cordova:copyLibs
(and update the plugin.xml if the version has changed.
also remove old libs if the version has changed).

This is a workaround for simply having an 'aar' dependency.
The 'a' alias from tools/bash.inc automates the deployment
of updated libs, so I can iterate quickly on the cordova plugin ...